pyproject.toml
[tool.poetry]
name = "random"
version = "0.1.0"
description = ""
authors = ["Osinachi Chukwujama  <Osinachi.Chukwujama@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bytewax = "^0.9.0"
kafka-python = "^2.0.2"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
